[Data] Normalize block types before internal multi-block operations (#43764)
Applying grouping operations on `Dataset`s with different underlying `Block` types can cause exceptions (e.g. various `AttributeError`s) due to `BlockAccessor`s assuming that all input block types are of the same type.

We handle this case by normalizing the blocks (either `ArrowBlock` or `PandasBlock`) to the first block type before applying the rest of the grouping/aggregation logic.
